+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleViewsTests.java
@@ -158,7 +158,7 @@ public class ModuleViewsTests {
     @Test
     public void onConfigurationChangedOrientationDisabled() {
         Countly mCountly = new Countly();
-        CountlyConfig config = (new CountlyConfig(getContext(), "appkey", "http://test.count.ly")).setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting();
+        CountlyConfig config = (new CountlyConfig(getContext(), "appkey", "http://test.count.ly")).setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting().setTrackOrientationChanges(false);
         mCountly.init(config);
         EventProvider ep = TestUtils.setEventProviderToMock(mCountly, mock(EventProvider.class));